dev: Add tests
